Hardware Class fixes
