--- MeepMeepTesting/src/main/java/com/example/meepmeeptesting/MeepMeepTesting.java
@@ -1,35 +1,0 @@
-package com.example.meepmeeptesting;
-
-import com.acmerobotics.roadrunner.geometry.Pose2d;
-import com.acmerobotics.roadrunner.geometry.Vector2d;
-import com.noahbres.meepmeep.MeepMeep;
-import com.noahbres.meepmeep.roadrunner.DefaultBotBuilder;
-import com.noahbres.meepmeep.roadrunner.entity.RoadRunnerBotEntity;
-
-public class MeepMeepTesting {
-    public static void main(String[] args) {
-        MeepMeep meepMeep = new MeepMeep(800);
-
-        RoadRunnerBotEntity myBot = new DefaultBotBuilder(meepMeep)
-                // Set bot constraints: maxVel, maxAccel, maxAngVel, maxAngAccel, track width
-                .setConstraints(46, 46, Math.toRadians(180), Math.toRadians(180), 12)
-                .followTrajectorySequence(drive ->
-                        drive.trajectorySequenceBuilder(new Pose2d(37, 65, Math.toRadians(90)))
-                                .setReversed(true)
-                                .splineTo(new Vector2d(35, 12), Math.toRadians(225))
-                                .addDisplacementMarker(() -> {})
-                                .waitSeconds(0.5)
-                                .setReversed(false)
-                                .forward(7.75)
-                                .turn(Math.toRadians(-45))
-                                .forward(20)
-                                .build()
-                );
-
-        meepMeep.setBackground(MeepMeep.Background.FIELD_POWERPLAY_OFFICIAL)
-                .setDarkMode(true)
-                .setBackgroundAlpha(0.95f)
-                .addEntity(myBot)
-                .start();
-    }
-}
